Fixed navigation and authentication
